Increase spacing when higher numbers on printing pascal triangle and return long to calculate factorial

+++ helpers.cpp
@@ -12,7 +12,7 @@ float cin_number() {
     return input;
 }
 
-int calc_factorial(float number) {
+long calc_factorial(float number) {
    int i, result = 1;
    for(i = number; i>1; i--)
       result *= i;
